Made parser cache use tokens instead of strings
